Change various keyword keys to strings in maps
Change various keyword keys to strings in the return values
of the outputs, template, tags, and transform functions.

Templates have a lot of keys that aren't valid keywords,
like "Fn::If". We previously used a mix of keywords and strings.
Now all keys are strings all the time, so users that may be
modifying the results don't have to look up which is which.
Users may still use any combination of keywords and strings in their
own code just as before.